(feat) move kraken to new location
