+++ hummingbot/connector/exchange/kraken/kraken_constants.py
