+++ hummingbot/connector/exchange/kraken/kraken_market.pxd
