+++ hummingbot/core/utils/market_mid_price.py
@@ -3,7 +3,7 @@ from decimal import Decimal
 from typing import Optional
 import cachetools.func
 from hummingbot.connector.exchange.binance.binance_market import BinanceMarket
-from hummingbot.market.kraken.kraken_market import KrakenMarket
+from hummingbot.connector.exchange.kraken.kraken_market import KrakenMarket
 
 
 BINANCE_PRICE_URL = "https://api.binance.com/api/v3/ticker/bookTicker"
