+++ setup.py
@@ -41,7 +41,7 @@ def main():
         "hummingbot.connector.exchange.coinbase_pro",
         "hummingbot.market.huobi",
         "hummingbot.connector.exchange.radar_relay",
-        "hummingbot.market.kraken",
+        "hummingbot.connector.exchange.kraken",
         "hummingbot.connector.exchange.crypto_com",
         "hummingbot.connector.exchange.kucoin",
         "hummingbot.script",
