+++ test/integration/test_kraken_market.py
@@ -4,8 +4,8 @@ from os import unlink
 from os.path import join, realpath
 import sys; sys.path.insert(0, realpath(join(__file__, "../../../")))
 
-from hummingbot.market.kraken.kraken_market import KrakenMarket
-from hummingbot.market.kraken.kraken_utils import convert_to_exchange_trading_pair
+from hummingbot.connector.exchange.kraken.kraken_market import KrakenMarket
+from hummingbot.connector.exchange.kraken.kraken_utils import convert_to_exchange_trading_pair
 from hummingbot.connector.markets_recorder import MarketsRecorder
 from hummingbot.model.sql_connection_manager import (
     SQLConnectionManager,
